backend/requirements.txt
Flask
Flask-CORS
tensorflow==2.12.0
numpy
Pillow
opencv-python
